Switch to conversion assertions.
